+++ changelog.md
@@ -5,6 +5,7 @@
 * Added PTG-3 based villager chat AI
     * Enable in Config
     * Speak their name once to start a conversation
+* Added more phrases
 
 # 7.3.21